gcp/requirements.txt
google-cloud-storage
tensorflow==2.7.4
Pillow==9.4.0